fix: show post.url if the post does not have a thumbnail
Right now, if a post does not have a post.thumbnail_url and is just a
link to another website, the link is not displayed... so the user cannot
go to the actual article.

This copies the link code from the Post object and provides a link if
there is a post.url but not a post.thumbnail_url.