fix: only list/search subscribed communities
Right now, listing and searching 'All' communities makes it difficult to
find the appropriate community to post to, particularly if it has a
common name such as technology, news, or linux.

Downside of this is that you can only post to communities you subscribe
to... but this was what we did before 9c9a3dc.