fix: parse date strings as UTC
For posts, comments, and other objects, the `published` attribute is a
string containing an ISO 8601 timestamp. This timestamp is UTC, though
it is not explicit in the string's contents. When the `Date` constructor
is used on such strings, they are parsed as though they are local
timestamps. According to [1]:

> When the time zone offset is absent, date-only forms are
> interpreted as a UTC time and date-time forms are interpreted as local
> time. This is due to a historical spec error that was not consistent
> with ISO 8601 but could not be changed due to web
> compatibility.

This patch explicitly appends a 'Z' to the `published` string when it is
used as the argument to the `Date` constructor.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date