fix: force navbar avatar to be square
Though the width and height attributes are set to 32 in this component,
the height is overridden by the more general styling for `<img>` tags,
which sets image heights to `auto`. For users with non-square avatars,
their avatar may overflow beyond its ring border. Specify the
`aspect-square` class to force a 1:1 aspect ratio. The Avatar component
used for posts, comments, etc. already does this.