feat: store sidebar expanded state in settings
Add an expandSidebar setting to remember if the sidebar is expanded or
collapsed.  This allows the user to refresh or open a post in a new tab
and keep the same sidebar state (rather than it always being expanded).

One possible quirk is that if a user has it collapsed in one tab, but
expanded in another, it will remember which ever state was set last
(most recently).  However, changing the state in one tab will not impact
another (until a refresh occurs).

Fix #68