misc: allow community card to be full width and remove left border (#72)
Right now, the community card model displays the information in a very
narrow column with a lot of extra whitespace on the side and a weird
left border.  Likewise, user pages also include a left border if the
StickyCard is placed at the top.

This removes the left border completely, and allows the StickyCard to be
wider in order to allow for more information to be displayed.