Remove redundant <hr> tags within expandable headers
There are <hr> tags within the expandable headers, the clickable components that expand or collapse when you click on them (for example, at the left sidebar of the home page on desktop) that seem redundant to me, as the expandable headers themselves are separated with more <hr> tags - so they look a bit cluttered. I'd like if you consider removing this <hr> tags placed within this expandable headers and giving said sidebars a cleaner look.